@@ -4,6 +4,7 @@ namespace App\Form;
 
 use App\Entity\Propositions;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
@@ -14,8 +15,12 @@ class PropositionsType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('firstname')
-            ->add('city')
+            ->add('firstname', TextType::class, [
+                'label' => 'Ton prénom'
+            ])
+            ->add('city', TextType::class, [
+        'label' => 'Ta ville '
+            ])
             ->add('date', DateType::class, [
                 'html5' => true,
                 'widget' => 'single_text',
